Entrega Final Terminado Completo Funcional Todos Los Clientes y Servidor

+++ Servidor JPA/SWServidor/SWServidor/src/main/java/model/ProductosPedidos.java
@@ -88,7 +88,7 @@ public class ProductosPedidos implements Serializable {
         this.fecha = fecha;
     }
 
-    public ProductosPedidos(long productoidProducto, BigInteger pedidosidPedido) {
+    public ProductosPedidos(long productoidProducto, long pedidosidPedido) {
         this.productosPedidosPK = new ProductosPedidosPK(productoidProducto, pedidosidPedido);
     }
